refactor!: replace spaces in parameter/event names
BREAKING CHANGE: This changes three aspects of the input and output
formats that are backwards incompatible:

1. The {{name}} and {{name_escaped}} input parameters have been removed,
   and {{name_enum}} renamed to {{event}}.
2. All spaces in output property names have been replaced with
   underscores. E.g. 'input action' => 'input_action', 'event format' =>
   'event_format'., 'active entry' => 'active_entry'
3. The default event_format is now:
   {"event":"{{event}}", "value":"{{value_escaped}}", "data":"{{data_escaped}}"}

I don't see the value in the old format. The space is harder to parse
and multiple formats only complicated documentation. I may reconsider
this move if this fork sees adoption, but I'll cross that bridge when I
get there.