feat!: emit ACTIVE_ENTRY when selection changes
BREAKING CHANGE: This relies on a patch to Rofi to call a
selection_changed mode function whenever the user changes the active
row (either by moving the focus, or by filtering).

And so, this changes the behavior of rofi-blocks' ACTIVE_ENTRY event, so
that it is emitted more often, and is not specifically tied to
CUSTOM_KEY (or COMPLETE, which I added in 6f5c027).

Ref: 6f5c027c8641