refactor!: rethink and rename COMPLETE event
BREAKING CHANGE: This renames COMPLETE_ENTRY to COMPLETE, and changes it
to deliver empty strings in its value and data fields. The purpose of
this event is a wildcard; it's not necessarily tied to the active entry,
so the _ENTRY suffix was misleading. It's not really tied to the current
input either, so it doesn't make sense to send that either.

So now it's an empty event, like CANCEL and EXIT. As a wildcard, it's up
to the user to exploit it however they like. The INPUT_CHANGE and
ACTIVE_ENTRY events can provide any state they may possible need anyhow.

Other changes:
- The data field for ACCEPT_INPUT and ACCEPT_INPUT_ALT events will now
  be "1" if an entry is active, and "" if not applicable (e.g. empty
  listview).
- Reformats the README to group events with similar fields.
- Adds more explicit datatypes to the README's Events table.