fix!: return (and use) pattern from mode_preprocess_input
BREAKING CHANGE: This relies on another Rofi patch, which changes how it
handles mode_preprocess_input's return value; treating it as a
substitute for the input text to filter lines against (without actually
changing the input bar). This allows users of rofi-blocks to set their
own static filter (or set it to "" to never filter anything), no matter
the contents of the input bar.