feat!: fix and enhance case_sensitive field
BREAKING CHANGE: Adds another Rofi patch to introduce a
rofi_view_set_case_sensitive function for changing case sensitivity
on-the-fly. This patch (and the one in 81ffac3) were necessary because
the Settings and RofiViewState structs, which house this (and other
desireable) state, are hidden behind opaque pointers. As the patches add
up, I'm starting to think it may be wiser to modify Rofi's meson.build
to include more of its headers, instead, but I'll let future-Henrik
worry about that.

Ref: 81ffac393ca9